+++ include/zephyr/drivers/watchdog.h
@@ -161,14 +161,6 @@ __subsystem struct wdt_driver_api {
  */
 __syscall int wdt_setup(const struct device *dev, uint8_t options);
 
-static inline int z_impl_wdt_setup(const struct device *dev, uint8_t options)
-{
-	const struct wdt_driver_api *api =
-		(const struct wdt_driver_api *)dev->api;
-
-	return api->setup(dev, options);
-}
-
 /**
  * @brief Disable watchdog instance.
  *
@@ -184,14 +176,6 @@ static inline int z_impl_wdt_setup(const struct device *dev, uint8_t options)
  * @retval -errno In case of any other failure.
  */
 __syscall int wdt_disable(const struct device *dev);
-
-static inline int z_impl_wdt_disable(const struct device *dev)
-{
-	const struct wdt_driver_api *api =
-		(const struct wdt_driver_api *)dev->api;
-
-	return api->disable(dev);
-}
 
 /**
  * @brief Install a new timeout.
@@ -239,20 +223,13 @@ static inline int wdt_install_timeout(const struct device *dev,
  */
 __syscall int wdt_feed(const struct device *dev, int channel_id);
 
-static inline int z_impl_wdt_feed(const struct device *dev, int channel_id)
-{
-	const struct wdt_driver_api *api =
-		(const struct wdt_driver_api *)dev->api;
-
-	return api->feed(dev, channel_id);
-}
-
 #ifdef __cplusplus
 }
 #endif
 
 /** @} */
 
+#include <zephyr/drivers/watchdog/internal/watchdog_impl.h>
 #include <zephyr/syscalls/watchdog.h>
 
 #endif /* ZEPHYR_INCLUDE_DRIVERS_WATCHDOG_H_ */
